+++ src/pages/welcome/styles.less
@@ -1,0 +1,49 @@
+/* WelcomePage.css */
+:root {
+  --primary-color: #1890ff;
+  --secondary-color: #e74c3c;
+  --tertiary-color: #9b59b6;
+  --background-gradient: linear-gradient(
+    45deg,
+    var(--primary-color),
+    var(--secondary-color),
+    var(--tertiary-color),
+    var(--primary-color)
+  );
+}
+
+.welcome-page {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background: var(--background-gradient);
+  background-size: 400% 400%;
+  animation: gradientAnimation 15s linear infinite;
+  color: white;
+  font-family: "Arial", sans-serif;
+}
+
+.welcome-heading {
+  font-size: 48px;
+  font-weight: bold;
+  margin-bottom: 20px;
+}
+
+.content-intro {
+  font-size: 24px;
+}
+
+/* Animation for Background Gradient */
+@keyframes gradientAnimation {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
