+++ src/utils/utils.ts
@@ -1,0 +1,3 @@
+export const getRandomInt = (min: number, max: number) => {
+  return Math.floor(Math.random() * (max - min)) + min;
+};
